Extract CommandRunner interface and provide default implementation
